requirements.txt
django==4.0.4
Markdown==3.1.1
django-filter==21.1
djangorestframework==3.13.1
Pillow==7.1.2
dnspython==2.2.1
djongo==1.3.6
pymongo==3.12.3
django-environ==0.8.1
sqlparse==0.2.4
psycopg2==2.9.3
boto3==1.23.6
django-storages==1.12.3
flake8==4.0.1
django-cors-headers==3.12.0
pyyaml==6.0
django-rosetta==0.9.8
django-parler==2.3
django-extensions==3.2.0
factory-boy==3.2.1
drf-nested-routers==0.93.4
black==22.6.0
isort==5.10.1
dj_database_url===1.0.0
